third_party/blink/web_tests/http/tests/websocket/alert-in-event-handler-expected.txt

ALERT: waiting for open
ALERT: message handled.first message
ALERT: message handled.second message
Make sure event handler called serially.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

alert will suspend/resume WebSocket.
onopen should fire later.
Connected
Enter onmessage: first message
Leave onmessage: first message
Enter onmessage: second message
Leave onmessage: second message
Closed
PASS successfullyParsed is true

TEST COMPLETE
